LPS-169073 Move the usage of abstract method getPortalCacheManagerConfiguration() in BasePortalCacheManager to EhcachePortalCacheManager to remove the abstract method from BasePortalCacheManager
